# go-server/api/swagger.yaml
openapi: 3.0.0
info:
  title: Moralis API
  version: "2"
servers:
- url: https://virtserver.swaggerhub.com/vit534/apispec/2
  description: SwaggerHub API Auto Mocking
- url: https://deep-index.moralis.io/api/v2
security:
- ApiKeyAuth: []
paths:
...
  /resolve/{domain}:
    get:
      tags:
      - resolve
      summary: Resolves an Unstoppable domain and returns the address
      description: |
        Resolves an Unstoppable domain and returns the address
      operationId: resolveDomain
      parameters:
      - name: currency
        in: query
        description: The currency to query
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: eth
          enum:
          - eth
          - 0x1
          default: eth
      - name: domain
        in: path
        description: Domain to be resolved
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: brad.crypto
      responses:
        "200":
          description: Returns an address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resolve'
        "404":
          description: Returns an address
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
      security:
      - ApiKeyAuth: []
  /resolve/{address}/reverse:
    get:
      tags:
      - resolve
      summary: Return the ENS domain when available (Only for ETH)
      description: |
        Resolves an ETH address and find the ENS name
      operationId: resolveAddress
      parameters:
      - name: address
        in: path
        description: The address to be resolved
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
      responses:
        "200":
          description: Returns an ENS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ens'
      security:
      - ApiKeyAuth: []
  /{pair_address}/reserves:
    get:
      tags:
      - defi
      summary: Get liquidity pair reserves for an Uniswap V2 based Exchange.
      description: Get the liquidity reserves for a given pair address. Only Uniswap
        V2 based exchanges supported at the moment.
      operationId: getPairReserves
      parameters:
      - name: chain
        in: query
        description: The chain to query
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/chainList'
      - name: to_block
        in: query
        description: To get the reserves at this block number
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: to_date
        in: query
        description: "Get the reserves to this date (any format that is accepted by\
          \ momentjs)\n* Provide the param 'to_block' or 'to_date'\n* If 'to_date'\
          \ and 'to_block' are provided, 'to_block' will be used.\n"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: provider_url
        in: query
        description: web3 provider url to user when using local dev chain
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: pair_address
        in: path
        description: Liquidity pair address
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0xa2107fa5b38d9bbd2c461d6edf11b11a50f6b974
      responses:
        "200":
          description: Returns the pair reserves
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reservesCollection'
      security:
      - ApiKeyAuth: []
  /{token0_address}/{token1_address}/pairAddress:
    get:
      tags:
      - defi
      summary: Get pair address based on token0 and token1 address
      description: |
        Fetches and returns pair data of the provided token0+token1 combination.
        The token0 and token1 options are interchangable (ie. there is no different outcome in "token0=WETH and token1=USDT" or "token0=USDT and token1=WETH")
      operationId: getPairAddress
      parameters:
      - name: chain
        in: query
        description: The chain to query
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/chainList'
      - name: to_block
        in: query
        description: To get the reserves at this block number
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: to_date
        in: query
        description: "Get the reserves to this date (any format that is accepted by\
          \ momentjs)\n* Provide the param 'to_block' or 'to_date'\n* If 'to_date'\
          \ and 'to_block' are provided, 'to_block' will be used.\n"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: exchange
        in: query
        description: The factory name or address of the token exchange
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: uniswapv2
          enum:
          - uniswapv2
          - uniswapv3
          - sushiswapv2
          - pancakeswapv2
          - pancakeswapv1
          - quickswap
      - name: token0_address
        in: path
        description: Token0 address
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0x2b591e99afe9f32eaa6214f7b7629768c40eeb39
      - name: token1_address
        in: path
        description: Token1 address
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0xdac17f958d2ee523a2206206994597c13d831ec7
      responses:
        "200":
          description: Returns the pair address of the two tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reservesCollection'
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
...
    chainList:
      type: string
      example: eth
      enum:
      - eth
      - 0x1
      - ropsten
      - 0x3
      - rinkeby
      - 0x4
      - goerli
      - 0x5
      - kovan
      - 0x2a
      - polygon
      - 0x89
      - mumbai
      - 0x13881
      - bsc
      - 0x38
      - bsc testnet
      - 0x61
      - avalanche
      - 0xa86a
      - avalanche testnet
      - 0xa869
      - fantom
      - 0xfa
      default: eth
...
    ens:
      required:
      - name
      properties:
        name:
          type: string
          description: Resolved ENS address
          example: Vitalik.eth
      example:
        name: Vitalik.eth
    resolve:
      required:
      - address
      properties:
        address:
          type: string
          description: Resolved domain address
          example: 0x057Ec652A4F150f7FF94f089A38008f49a0DF88e
      example:
        address: 0x057Ec652A4F150f7FF94f089A38008f49a0DF88e
    reservesCollection:
      required:
      - reserve0
      - reserve1
      properties:
        reserve0:
          type: string
          description: reserve0
          example: "1177323085102288091856004"
        reserve1:
          type: string
          description: reserve1
          example: "9424175928981149993184"
      example:
        reserve1: "9424175928981149993184"
        reserve0: "1177323085102288091856004"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-Key
      in: header